Migrate hyco-websocket Tape tests to Jest framework
